renames type.token to type.jwt
adds better test for jwt type
