+++ src/types/index.ts
@@ -2,19 +2,19 @@
 import email from './email';
 import uuid from './uuid';
 import uri from './uri';
-import token from './token';
+import jwt from './jwt';
 import phone from './phone';
 
 export { email };
 export { uuid };
 export { uri };
-export { token };
+export { jwt };
 export { phone };
 
 export default {
   email,
   uuid,
   uri,
-  token,
+  jwt,
   phone,
 };
